Move TopSpin reader to NMR plugin
