+++ environments/environment_cp.yml
@@ -33,7 +33,6 @@ dependencies:
     - matplotlib-base
     - numpy
     - numpydoc
-    - quaternion>=2024.0.7
     - osqp
     - pluggy
     - pint
