+++ plugins/spectrochempy-nmr/.github/workflows/test.yml
@@ -1,0 +1,22 @@
+name: Test spectrochempy-topspin
+
+on: [push, pull_request]
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.11", "3.12", "3.13"]
+
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Install core spectrochempy
+        run: pip install -e ../..
+      - name: Install plugin
+        run: pip install -e .
+      - name: Run tests
+        run: python -m pytest tests/ -v
