+++ plugins/spectrochempy-nmr/NMRGLUE_LICENSE.rst
