+++ plugins/spectrochempy-nmr/pyproject.toml
@@ -1,0 +1,23 @@
+[build-system]
+requires = ["setuptools>=64"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "spectrochempy-nmr"
+version = "0.1.0"
+description = "NMR readers and tools plugin for SpectroChemPy"
+requires-python = ">=3.11"
+dependencies = [
+    "spectrochempy>=0.8,<0.9",
+    "numpy",
+    "numpy-quaternion>=2024.0.7",
+]
+
+[project.entry-points."spectrochempy.plugins"]
+nmr = "spectrochempy_nmr:NMRPlugin"
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
