+++ plugins/spectrochempy-nmr/src/spectrochempy_nmr/nmrglue.py
