+++ requirements/requirements.txt
@@ -25,7 +25,6 @@ lazy_loader
 matplotlib
 numpy
 numpydoc
-numpy-quaternion >= 2024.0.7
 osqp
 pluggy
 pint
