+++ requirements/requirements_nmr.txt
@@ -21,4 +21,4 @@
 
 # NMR dependencies
 # ----------------
-spectrochempy[quaternion]
+spectrochempy-nmr>=0.1,<0.2
